requirements.txt
gensim==3.8.3
Keras==2.2.4
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
scikit-learn==0.24.2
tensorflow-gpu==1.14.0
